5a. building frontend UI pt.1
created Navbar, Home page, Create page, Product card, product.js

+++ backend/controllers/productControllers.js
@@ -34,9 +34,10 @@ export const updateProduct = async (req, res) => {
     const product = req.body;
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
-        console.log("Product not found")
-        return res.status(404).json({ success: false, message: "Product not found" });
+        console.log("Invalid product ID")
+        return res.status(404).json({ success: false, message: "Invalid product ID" });
     }
+
     try {
         const updatedProduct = await Product.findByIdAndUpdate(id, product, { new: true });
         res.status(200).json({ success: true, data: updatedProduct });
@@ -49,11 +50,16 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        console.log("Invalid product ID")
+        return res.status(404).json({ success: false, message: "Invalid product ID" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
     } catch (error) {
         console.log("Error in deleting product: ", error.message);
-        res.status(400).json({ success: false, message: "Product does not exist"})
+        res.status(500).json({ success: false, message: "Server error" })
     }
 };
